% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockermachine.R
\docType{package}
\name{dockermachine-package}
\alias{dockermachine}
\alias{dockermachine-package}
\title{dockermachine}
\description{
An R interface to docker-machine.
}
\references{
\url{https://docs.docker.com/machine/reference}
}
\author{
\strong{Maintainer}: Carl Boettiger \email{cboettig@gmail.com} (http://orcid.org/0000-0002-1642-628X) [copyright holder]

}
