% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{dockermachine_cmd}
\alias{dockermachine_cmd}
\alias{dockermachine_version}
\title{dockermachine_cmd}
\usage{
dockermachine_cmd(...)

dockermachine_version()
}
\arguments{
\item{...}{Arguments to be passed to \code{system2('docker-machine', ...)}, e.g.
\code{new_content(path)} is basically \code{dockermachine_cmd(c('new', path))} (i.e.
run the command \command{docker-machine new path}).}
}
\description{
Run an arbitrary docker-machine command
}
\section{Functions}{
\itemize{
\item \code{dockermachine_cmd}: Run an arbitrary docker-machine command.

\item \code{dockermachine_version}: Return the version number of docker-machine if possible, which is
extracted from the output of \code{dockermachine_cmd('version')}.
}}

