% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_dockermachine}
\alias{install_dockermachine}
\alias{update_dockermachine}
\title{Install dockermachine}
\usage{
install_dockermachine(version = "latest", use_brew = Sys.which("brew") !=
  "", force = FALSE)

update_dockermachine()
}
\arguments{
\item{version}{The dockermachine version number, e.g., \code{0.26}; the special value
\code{latest} means the latest version (fetched from Github releases).}

\item{force}{Whether to install dockermachine even if it has already been installed.
This may be useful when upgrading dockermachine}
}
\description{
Download the appropriate dockermachine executable for your platform from Github and
try to copy it to a system directory so \pkg{blogdown} can run the
\command{dockermachine} command to build a site. \code{update_dockermachine()} is a wrapper of
\code{install_dockermachine(force = TRUE)}.
}
\details{
This function tries to install dockermachine to \code{Sys.getenv('APPDATA')} on
Windows, \file{~/Library/Application Support} on macOS, and \file{~/bin/} on
other platforms (such as Linux). If these directories are not writable, the
package directory \file{dockermachine} of \pkg{dockermachine} will be used. If it still
fails, you have to install dockermachine by yourself and make sure it can be found via
the environment variable \code{PATH}.

This is just a helper function and may fail to choose the correct dockermachine
executable for your operating system, especially if you are not on Windows or
Mac or a major Linux distribution. When in doubt, read the docker-machine documentation
and install it by yourself: \url{https://docs.docker.com/machine/}.
}
