% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_machine}
\alias{install_machine}
\alias{update_machine}
\title{Install docker-machine}
\usage{
install_machine(version = "latest", force = FALSE)

update_machine()
}
\arguments{
\item{version}{The docker-machine version number, e.g., \code{0.12.2}; the special value
\code{latest} means the latest version (fetched from Github releases).}

\item{force}{Whether to install docker-machine even if it has already been installed.
This may be useful when upgrading docker-machine}
}
\description{
Download the appropriate docker-machine executable for your platform from Github and
try to copy it to a system directory. \code{update_machine()} is a wrapper of
\code{install_machine(force = TRUE)}.
}
\details{
This function tries to install docker-machine to \code{Sys.getenv('APPDATA')} on
Windows, \file{~/Library/Application Support} on macOS, and \file{~/bin/} on
other platforms (such as Linux). If these directories are not writable, the
package directory \file{dockermachine} of \pkg{dockermachine} will be used. If it still
fails, you have to install docker-machine by yourself and make sure it can be found via
the environment variable \code{PATH}.

This is just a helper function and may fail to choose the correct docker-machine
executable for your operating system, especially if you are not on Windows or
Mac or a major Linux distribution. When in doubt, read the docker-machine documentation
and install it by yourself: \url{https://docs.docker.com/machine/}.
}
