% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{machine_cmd}
\alias{machine_cmd}
\alias{machine_version}
\title{machine_cmd}
\usage{
machine_cmd(...)

machine_version()
}
\arguments{
\item{...}{Arguments to be passed to \code{system2('docker-machine', ...)}, e.g.
\code{new_content(path)} is basically \code{machine_cmd(c('new', path))} (i.e.
run the command \command{docker-machine new path}).}
}
\description{
Run an arbitrary docker-machine command
}
\section{Functions}{
\itemize{
\item \code{machine_cmd}: Run an arbitrary docker-machine command.

\item \code{machine_version}: Return the version number of docker-machine if possible, which is
extracted from the output of \code{machine_cmd('version')}.
}}

