% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockermachine.R
\name{machine_create}
\alias{machine_create}
\title{docker-machine create}
\usage{
machine_create(driver = c("none", "virtualbox", "amazonec2", "digitalocean",
  "exoscale", "generic", "google", "softlayer", "azure", "hyperv", "openstack",
  "rackspace", "vmwarefusion", "vmwarevcloudair", "vmwarevsphere"),
  name = getOption("DOCKERMACHINE_NAME", "machine"), driver_conf = list(),
  engine_conf = list(install_url = "https://get.docker.com", opt =
  character(), insecure_registry = character(), regsitry_mirror = character(),
  label = character(), storage_driver = character(), env = character()),
  swarm = FALSE, swarm_conf = list(image = "swarm:latest", master =
  character(), discovery = character(), strategy = "spread", opt = character(),
  host = "tcp://0.0.0.0:3376", addr = character(), experimental = character()),
  help = FALSE, env = character(), ...)
}
\arguments{
\item{driver}{name of the driver (cloud platform) to create machine with}

\item{name}{the machine name, default is 'machine', but can also be
set using \code{options("DOCKERMACHINE_NAME" = "<PICK-A-NAME>")}}

\item{driver_conf}{list of additional arguments for docker driver config,
see \code{machine_help("drivername")} for driver-specific arguments.}

\item{engine_conf}{list of additional arguments for docker driver,
see details}

\item{swarm}{logical, default FALSE. Display the Swarm config instead
of the Docker daemon}

\item{swarm_conf}{list of additional arguments for docker swarm config,
see details}

\item{help}{logical, default FALSE. Set to TRUE to display help
information for the command instead of executing the command.}

\item{env}{Additional character vector of name=value strings to set
environment variables.}

\item{...}{Additional arguments to \code{\link{system2}}}
}
\value{
If stdout = TRUE or stderr = TRUE, a character vector giving
 the output of the command, one line per character string. (Output
 lines of more than 8095 bytes will be split.) If the command could
 not be run an R error is generated. If command runs but gives a
 non-zero exit status this will be reported with a warning and in
 the attribute "status" of the result: an attribute "errmsg" may
 also be available.

 In other cases, the return value is an error code (0 for success),
 given the invisible attribute (so needs to be printed explicitly).
 If the command could not be run for any reason, the value is 127.
 Otherwise if wait = TRUE the value is the exit status returned by
 the command, and if wait = FALSE it is 0 (the conventional success
 value).
}
\description{
Create a machine. Requires the --driver flag to indicate which provider
(VirtualBox, DigitalOcean, AWS, etc.) the machine should be created on,
and an argument to indicate the name of the created machine.
}
\examples{
\dontrun{
machine_create("virtualbox")
}

}
\references{
\url{https://docs.docker.com/machine/reference/create/}
 @details
--driver, -d "none"                                                                                  Driver to create machine with.
--engine-install-url "https://get.docker.com"                                                        Custom URL to use for engine installation [$MACHINE_DOCKER_INSTALL_URL]
--engine-opt [--engine-opt option --engine-opt option]                                               Specify arbitrary flags to include with the created engine in the form flag=value
--engine-insecure-registry [--engine-insecure-registry option --engine-insecure-registry option]     Specify insecure registries to allow with the created engine
--engine-registry-mirror [--engine-registry-mirror option --engine-registry-mirror option]           Specify registry mirrors to use [$ENGINE_REGISTRY_MIRROR]
--engine-label [--engine-label option --engine-label option]                                         Specify labels for the created engine
--engine-storage-driver                                                                              Specify a storage driver to use with the engine
--engine-env [--engine-env option --engine-env option]                                               Specify environment variables to set in the engine
--swarm                                                                                              Configure Machine with Swarm
--swarm-image "swarm:latest"                                                                         Specify Docker image to use for Swarm [$MACHINE_SWARM_IMAGE]
--swarm-master                                                                                       Configure Machine to be a Swarm master
--swarm-discovery                                                                                    Discovery service to use with Swarm
--swarm-strategy "spread"                                                                            Define a default scheduling strategy for Swarm
--swarm-opt [--swarm-opt option --swarm-opt option]                                                  Define arbitrary flags for swarm
--swarm-host "tcp://0.0.0.0:3376"                                                                    ip/socket to listen on for Swarm master
--swarm-addr                                                                                         addr to advertise for Swarm (default: detect and use the machine IP)
--swarm-experimental
}
